% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellNetVis_UMAP.R
\name{cell_cell_format_cellchat}
\alias{cell_cell_format_cellchat}
\title{Restructure output of CellChat}
\usage{
cell_cell_format_cellchat(cell_chat_out, receptor, ligand)
}
\arguments{
\item{cell_chat_out}{Output from CellChat. The object is the out put of the "subsetCommunication" function}

\item{receptor}{gene name of the receptor}

\item{ligand}{gene name of the ligand}
}
\description{
This function formats CellChat data into a simplified cell-cell connection dataframe
}
\examples{
cell_cell_format_cellchat()

}
\keyword{UMAP}
