% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellNetVis_UMAP.R
\name{cell_cell_format_cpdb}
\alias{cell_cell_format_cpdb}
\title{Restructure output of cellphonedb}
\usage{
cell_cell_format_cpdb(cpdb_out, receptor, ligand)
}
\arguments{
\item{cpdb_out}{Output from cellphonedb. The object named something like "statistical_analysis_significant_means.txt"}

\item{receptor}{gene name of the receptor}

\item{ligand}{gene name of the ligand}
}
\description{
This function formats cpdb data into a simplified cell-cell connection dataframe
}
\examples{
cell_cell_format_cpdb()

}
\keyword{UMAP}
