% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellNetVis_UMAP.R
\name{plot_gene_UMAP_color_bycelltype}
\alias{plot_gene_UMAP_color_bycelltype}
\title{A Cat Function}
\usage{
plot_gene_UMAP_color_bycelltype(
  seurat_object,
  cell_type_col,
  cell_cell_connections,
  ligand_cell_type = NA,
  self_interactions = F,
  label_cell_type = T
)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
cat_function()

}
\keyword{cats}
