% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellNetVis_UMAP.R
\name{plot_gene_UMAP_exp_colored}
\alias{plot_gene_UMAP_exp_colored}
\title{Visualisation to colour the UMAP points by expression of two genes}
\usage{
plot_gene_UMAP_exp_colored(
  seurat_object,
  cell_type_col,
  cell_cell_connections,
  ligand_cell_type = NA,
  self_interactions = F,
  label_cell_type = T,
  receptor,
  ligand,
  percentile = 0.8
)
}
\arguments{
\item{seurat_object}{object with UMAP coordinates}

\item{cell_type_col}{column name of cell type}

\item{cell_cell_connections}{data.frame structured like the example “cell_cell_connections”.}

\item{ligand_cell_type}{do you want to limit the curves drawn to only starting at one cell type of interest? For example only cDC2 cells. Defaults to NA}

\item{self_interactions}{do you want to see loops for when cells are interacting with themselves? Defaults to F}

\item{label_cell_type}{do you want text labels for the cell types? Defaults to T}

\item{receptor}{gene name of the receptor (or any gene).}

\item{ligand}{gene name of the ligand (or any gene)}

\item{percentile}{the percentile cut off to call a gene as “highly expressed” and color a cell as having that gene “on”. Defaults to 0.8}
}
\description{
Colour UMAP by expression of two genes. The gene don't have to be receptors and ligands just a typical use case.
}
\examples{
plot_gene_UMAP_exp_colored()
}
\keyword{UMAP}
