% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellNetVis_UMAP.R
\name{plot_gene_UMAP_color_bycelltype}
\alias{plot_gene_UMAP_color_bycelltype}
\title{Visualisation to colour the UMAP points by cell type label}
\usage{
plot_gene_UMAP_color_bycelltype(
  seurat_object,
  cell_type_col,
  cell_cell_connections,
  ligand_cell_type = NA,
  self_interactions = F,
  label_cell_type = T
)
}
\arguments{
\item{seurat_object}{object with UMAP coordinates}

\item{cell_type_col}{column name of cell type}

\item{cell_cell_connections}{data.frame structured like the example “cell_cell_connections”.}

\item{ligand_cell_type}{do you want to limit the curves drawn to only starting at one cell type of interest? For example only cDC2 cells. Defaults to NA}

\item{self_interactions}{do you want to see loops for when cells are interacting with themselves? Defaults to F}

\item{label_cell_type}{do you want text labels for the cell types? Defaults to T}
}
\description{
Colour UMAP by Cell Type.
}
\examples{
plot_gene_UMAP_color_bycelltype()

}
\keyword{UMAP}
